Fix 'handleSpecialAction' undefined error in StrategyService
Resolved an issue where the 'handleSpecialAction' method of the StrategyService class was being called on an undefined context. This was due to the improper use of the StrategyService instance within the 'onStep' method in bot.js.

Changes include:
- Modified the export of the StrategyService class to export the class itself instead of an instance.
- Updated bot.js to use StrategyService.getInstance() for accessing the singleton instance of StrategyService, ensuring that the correct context is used when calling its methods.

These changes ensure that the StrategyService's methods are called with the appropriate 'this' context, fixing the TypeError related to undefined 'handleSpecialAction'.